// generated/models/ApiProposalFrameRequest.ts
// @ts-nocheck
/**
 * 6529.io API
 * JSON REST API for 6529.io. New here? A terminology overview and a step-by-step authentication walkthrough (guides &amp; auth) live at <a href=\"https://6529.io/tools/api\">https://6529.io/tools/api</a>. The raw machine-readable spec is downloadable at <a href=\"/openapi.yaml\">/openapi.yaml</a> and <a href=\"/openapi.json\">/openapi.json</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ApiProposalFrameRequest {
    /**
    * Owned drop upload URL or recognized IPFS, IPNS, or Arweave artwork URL
    */
    'media_url': string;
    'mime_type': ApiProposalFrameRequestMimeTypeEnum;
    'title': string;
    'layout': ApiProposalFrameRequestLayoutEnum;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "media_url",
            "baseName": "media_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "mime_type",
            "baseName": "mime_type",
            "type": "ApiProposalFrameRequestMimeTypeEnum",
            "format": ""
        },
        {
            "name": "title",
            "baseName": "title",
            "type": "string",
            "format": ""
        },
        {
            "name": "layout",
            "baseName": "layout",
            "type": "ApiProposalFrameRequestLayoutEnum",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiProposalFrameRequest.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum ApiProposalFrameRequestMimeTypeEnum {
    ImagePng = 'image/png',
    ImageJpeg = 'image/jpeg',
    ImageJpg = 'image/jpg',
    ImageGif = 'image/gif',
    ImageWebp = 'image/webp',
    VideoMp4 = 'video/mp4',
    VideoXMsvideo = 'video/x-msvideo',
    VideoQuicktime = 'video/quicktime',
    TextHtml = 'text/html'
}
export enum ApiProposalFrameRequestLayoutEnum {
    Portrait = 'portrait',
    Landscape = 'landscape'
}